+++ Sources/UICore/UI/Style/style_impl.cpp
@@ -103,13 +103,13 @@ namespace uicore
 	{
 		for (size_t i = 0; i < value_array.size(); i++)
 		{
-			set_value(name + "[" + Text::to_string(i) + "]", value_array[i]);
+			set_value(name + "[" + Text::to_string((int)i) + "]", value_array[i]);
 		}
 
 		size_t i = value_array.size();
 		while (true)
 		{
-			auto index_name = name + "[" + Text::to_string(i) + "]";
+			auto index_name = name + "[" + Text::to_string((int)i) + "]";
 			if (prop_type.find(index_name) == prop_type.end())
 				break;
 			set_value(index_name, StyleSetValue());
